+++ flask_monitoringdashboard/database/function_calls.py
@@ -46,6 +46,30 @@ def get_times():
         return result
 
 
+def get_hits(date_from=None):
+    """ Return all entries of measurements with the number of execution times which are called after 'date_from'
+    :param date_from: A datetime-object
+    """
+    with session_scope() as db_session:
+        result = db_session.query(FunctionCall.endpoint,
+                                  func.count(FunctionCall.execution_time).label('count')). \
+            filter(FunctionCall.time > date_from).group_by(FunctionCall.endpoint).all()
+        db_session.expunge_all()
+        return result
+
+
+def get_average(date_from=None):
+    """ Return the average of execution times which are called after 'date_from' grouped per endpoint
+    :param date_from: A datetime-object
+    """
+    with session_scope() as db_session:
+        result = db_session.query(FunctionCall.endpoint,
+                                  func.avg(FunctionCall.execution_time).label('average')). \
+            filter(FunctionCall.time > date_from).group_by(FunctionCall.endpoint).all()
+        db_session.expunge_all()
+        return result
+
+
 def get_data_between(time_from, time_to):
     """
         Returns all data in the FunctionCall table, for the export data option.
@@ -67,10 +91,22 @@ def get_data():
 
 
 def get_data_per_version(version):
-    """ Returns all data in the FuctionCall table, grouped by their version. """
+    """ Returns all data in the FunctionCall table, grouped by their version. """
     with session_scope() as db_session:
         result = db_session.query(FunctionCall.execution_time, FunctionCall.version). \
             filter(FunctionCall.version == version).all()
+        db_session.expunge_all()
+        return result
+
+
+def get_hits_per_version(version):
+    """ Returns the hits per endpoint per version """
+    with session_scope() as db_session:
+        result = db_session.query(FunctionCall.endpoint,
+                                  FunctionCall.version,
+                                  func.count(FunctionCall.endpoint).label('count')). \
+                    filter(FunctionCall.version == version). \
+                    group_by(FunctionCall.endpoint).all()
         db_session.expunge_all()
         return result