+++ CHANGELOG.rst
@@ -9,6 +9,10 @@ Unreleased
 ----------
 Changed
 
+- Added new graph: Version usage
+
+- Added column (Hits in past 7 days) in Measurements Overview
+
 - Fixed bug with configuration
 
 - Changed rows and column in outlier-table